requirements.txt
discord.py
Pillow
python-dotenv
APScheduler